+++ frontend/packages/data-portal/app/components/Breadcrumbs.tsx
@@ -32,7 +32,10 @@ function Breadcrumb({
   return link ? (
     <Link
       to={link}
-      className={cns(className, 'hover:text-sds-color-primitive-blue-400')}
+      className={cns(
+        className,
+        'hover:text-light-sds-color-primitive-blue-500',
+      )}
       onClick={() => {
         if (type) {
           plausible(Events.ClickBreadcrumb, {
@@ -116,12 +119,14 @@ export function Breadcrumbs({
 
   return (
     <div
-      className={`flex flex-col flex-auto gap-1 ${variant === 'neuroglancer' && 'max-w-xl'}`}
+      className={`flex flex-col flex-auto gap-1 ${
+        variant === 'neuroglancer' && 'max-w-xl'
+      }`}
       data-testid={TestIds.Breadcrumbs}
     >
       {returnToDepositionLink && (
         <Link
-          className="uppercase font-semibold text-sds-caps-xxxs leading-sds-caps-xxxs text-sds-color-primitive-blue-400"
+          className="uppercase font-semibold text-sds-caps-xxxs-600-wide leading-sds-caps-xxxs text-light-sds-color-primitive-blue-500"
           to={returnToDepositionLink}
           onClick={() =>
             plausible(Events.ClickBreadcrumb, {
@@ -138,11 +143,17 @@ export function Breadcrumbs({
       )}
 
       <div
-        className={`flex flex-row gap-sds-s 
-        text-sds-body-s leading-sds-body-s 
-        ${typeStyles[type]}
-        ${variant === 'neuroglancer' ? 'text-[13px] font-normal' : ''}
-        items-center whitespace-nowrap content-start`}
+        className={cns(
+          'flex',
+          'flex-row gap-sds-s',
+          typeStyles[type],
+          variant === 'neuroglancer' ? 'text-[13px] font-normal' : '',
+          'text-sds-body-s-400-wide',
+          'leading-sds-body-s',
+          'text-light-sds-color-primitive-gray-900',
+          'items-center whitespace-nowrap',
+          'content-start',
+        )}
       >
         <Breadcrumb
           text={