+++ frontend/packages/data-portal/app/components/Dataset/RunsTable.tsx
@@ -28,7 +28,7 @@ import { useI18n } from 'app/hooks/useI18n'
 import { useIsLoading } from 'app/hooks/useIsLoading'
 import { Run } from 'app/types/gql/datasetPageTypes'
 import { cnsNoMerge } from 'app/utils/cns'
-import { isDefined } from 'app/utils/nullish'
+import { setObjectNameAndGroundTruthStatus } from 'app/utils/setObjectNameAndGroundTruthStatus'
 import { inQualityScoreRange } from 'app/utils/tiltSeries'
 import { carryOverFilterParams, createUrl } from 'app/utils/url'
 
@@ -124,9 +124,9 @@ export function RunsTable() {
                 ) : (
                   <TableLink
                     className={cnsNoMerge(
-                      'text-sds-body-m leading-sds-body-m font-semibold text-sds-color-primitive-blue-400',
+                      'text-sds-body-m-400-wide leading-sds-body-m font-semibold text-light-sds-color-primitive-blue-500',
                       !isHoveringOverInteractable &&
-                        'group-hover:text-sds-color-primitive-blue-500',
+                        'group-hover:text-light-sds-color-primitive-blue-600',
                     )}
                     to={runUrl}
                   >
@@ -134,7 +134,7 @@ export function RunsTable() {
                   </TableLink>
                 )}
 
-                <p className="text-sds-body-xxs leading-sds-body-xxs text-sds-color-semantic-text-base-primary">
+                <p className="text-sds-body-xxs-400-wide leading-sds-body-xxs text-light-sds-color-semantic-base-text-primary">
                   {isLoadingDebounced ? (
                     <Skeleton className="max-w-[120px]" variant="text" />
                   ) : (
@@ -179,9 +179,15 @@ export function RunsTable() {
 
       columnHelper.accessor(
         (run) =>
-          run.annotationsAggregate?.aggregate
-            ?.map((aggregate) => aggregate.groupBy?.objectName)
-            .filter(isDefined) ?? [],
+          run.annotationsAggregate?.aggregate?.reduce((acc, aggregate) => {
+            const objectName = aggregate.groupBy?.objectName
+            const groundTruthStatus = !!aggregate.groupBy?.groundTruthStatus
+            return setObjectNameAndGroundTruthStatus(
+              objectName,
+              groundTruthStatus,
+              acc,
+            )
+          }, new Map<string, boolean>()) || new Map<string, boolean>(),
         {
           id: 'annotatedObjects',
 
@@ -196,7 +202,7 @@ export function RunsTable() {
               renderLoadingSkeleton={false}
               width={RunTableWidths.annotatedObjects}
             >
-              {getValue().length === 0 ? (
+              {getValue().size === 0 ? (
                 '--'
               ) : (
                 <AnnotatedObjectsList
@@ -221,11 +227,9 @@ export function RunsTable() {
                 buttonProps={{
                   sdsType: 'primary',
                   sdsStyle: 'minimal',
-                  startIcon: (
-                    <Icon sdsIcon="Cube" sdsType="button" sdsSize="s" />
-                  ),
+                  startIcon: <Icon sdsIcon="Cube" sdsSize="s" />,
                   className:
-                    '!min-w-[141px] min-h-[32px] hover:!bg-sds-color-primitive-gray-200 rounded-md',
+                    '!min-w-[141px] min-h-[32px] hover:!bg-light-sds-color-primitive-gray-200 rounded-md',
                 }}
                 event={{
                   datasetId: dataset.id,