+++ frontend/packages/data-portal/app/components/DepositionFilterBanner.tsx
@@ -30,36 +30,43 @@ export function DepositionFilterBanner({
   const { t } = useI18n()
 
   return (
-    <Callout className="!w-full" classes={{ message: 'w-full' }} intent="info">
-      <div className="flex w-full items-center gap-sds-l justify-between">
-        {/* TODO: (kne42) sync with design on what we want to do on overflow */}
-        <p className="text-sds-body-xs leading-sds-body-xs flex-1 overflow-hidden text-ellipsis whitespace-nowrap">
-          <I18n
-            i18nKey={labelI18n}
-            values={{
-              ...deposition,
-              url: `/depositions/${deposition.id}?${previousSingleDepositionParams}`,
+    <Callout
+      className="!w-full"
+      classes={{ message: 'w-full', root: '!items-center' }}
+      intent="info"
+      body={
+        <div className="flex w-full items-center gap-sds-l justify-between">
+          {/* TODO: (kne42) sync with design on what we want to do on overflow */}
+          <p className="text-sds-body-xs-400-wide leading-sds-body-xs flex-1 overflow-hidden text-ellipsis whitespace-nowrap">
+            <I18n
+              i18nKey={labelI18n}
+              values={{
+                ...deposition,
+                url: `/depositions/${deposition.id}?${previousSingleDepositionParams}`,
+              }}
+              tOptions={{ interpolation: { escapeValue: false } }}
+            />
+          </p>
+
+          <Button
+            onClick={() => {
+              setDepositionId(null)
+
+              const nextParams = new URLSearchParams(
+                previousSingleDatasetParams,
+              )
+              nextParams.delete(QueryParams.DepositionId)
+              nextParams.sort()
+              setPreviousSingleDatasetParams(nextParams.toString())
             }}
-            tOptions={{ interpolation: { escapeValue: false } }}
-          />
-        </p>
-
-        <Button
-          onClick={() => {
-            setDepositionId(null)
-
-            const nextParams = new URLSearchParams(previousSingleDatasetParams)
-            nextParams.delete(QueryParams.DepositionId)
-            nextParams.sort()
-            setPreviousSingleDatasetParams(nextParams.toString())
-          }}
-          sdsStyle="minimal"
-          sdsType="secondary"
-          className="shrink-0"
-        >
-          {t('removeFilter')}
-        </Button>
-      </div>
-    </Callout>
+            sdsStyle="minimal"
+            sdsType="secondary"
+            className="shrink-0"
+          >
+            {t('removeFilter')}
+          </Button>
+        </div>
+      }
+    />
   )
 }