+++ frontend/packages/data-portal/app/components/TablePageLayout/TableHeaderDefinition.tsx
@@ -21,18 +21,18 @@ export function TableHeaderDefinition({
 
   return (
     <div className="pt-sds-s px-sds-xl max-w-content">
-      <h1 className="text-sds-header-xxl leading-sds-header-xxl tracking-sds-header-xxl font-semibold">
+      <h1 className="text-sds-header-xxl-600-wide leading-sds-header-xxl tracking-sds-header-xxl-600-wide font-semibold">
         {title}
       </h1>
 
-      <p className="text-sds-body-s leading-sds-body-s tracking-sds-body-s max-w-[600px] mt-sds-s">
-        <span className="text-sds-color-semantic-text-base-secondary">
+      <p className="text-sds-body-s-400-wide leading-sds-body-s tracking-sds-body-s-400-wide max-w-[600px] mt-sds-s">
+        <span className="text-light-sds-color-semantic-base-text-secondary">
           {description}
         </span>
 
         {learnMoreLink && (
           <Link
-            className="text-sds-color-semantic-text-action-default hover:text-sds-color-semantic-text-action-hover"
+            className="text-light-sds-color-semantic-accent-text-action hover:text-light-sds-color-semantic-accent-text-action-hover"
             to={learnMoreLink}
           >
             {' '}