+++ frontend/packages/data-portal/app/components/ViewTomogramButton.tsx
@@ -47,7 +47,8 @@ export function ViewTomogramButton({
             <Link
               to={t('neuroglancerTutorialLink')}
               variant="dashed-underlined"
-              className="!text-sds-color-primitive-gray-300 text-sds-body-xxs !border-sds-color-primitive-gray-800"
+              stopPropagation
+              className="!text-light-sds-color-primitive-gray-300 text-sds-body-xxs-400-wide !border-light-sds-color-primitive-gray-800"
             >
               <p>
                 <I18n i18nKey="viewNeuroglancerTutorial" />
@@ -88,7 +89,7 @@ export function ViewTomogramButton({
           }
           disabled={!enabled}
           LinkComponent={Link}
-          {...buttonProps}
+          {...(buttonProps as ButtonProps)}
         >
           <span>{t('viewTomogram')}</span>
         </Button>