+++ frontend/packages/data-portal/app/constants/siteLinks.ts
@@ -1,8 +1,10 @@
-export enum SiteLinks {
+export enum SITE_LINKS {
   HOME = '/',
   COMPETITION = '/competition',
   BROWSE_DATA_DATASETS = '/browse-data/datasets',
   FAQ = 'https://chanzuckerberg.github.io/cryoet-data-portal/stable/cryoet_data_portal_docsite_faq.html',
   PRIVACY = '/privacy',
   DATA_SUBMISSION_POLICY = '/data-submission-policy',
+  DOCUMENTATION = 'https://chanzuckerberg.github.io/cryoet-data-portal',
+  GROUND_TRUTH_FLAG_DOCS = 'https://chanzuckerberg.github.io/cryoet-data-portal/stable/cryoet_data_portal_docsite_data.html#ground-truth-flagn',
 }